fix:main.dartを分解
